[Add] #174 - ReadStatus 새로운 버전에 맞게 추가
