+++ WSSiOS/WSSiOS/Source/Data/Base/TrashReadStatus.swift
@@ -40,43 +40,3 @@ enum TrashReadStatus: String, CaseIterable {
         }
     }
 }
-
-
-//
-//  ReadStatus.swift
-//  WSSiOS
-//
-//  Created by 이윤학 on 1/17/24.
-//
-
-import UIKit
-
-enum ReadStatus: String, CaseIterable {
-    case watched
-    case watching
-    case quit
-    
-    var tagImage: UIImage {
-        switch self {
-        case .watched: return .icTagFinished
-        case .watching: return .icTagReading
-        case .quit: return .icTagStop
-        }
-    }
-    
-    var tagText: String {
-        switch self {
-        case .watched: return "봤어요"
-        case .watching: return "보는 중"
-        case .quit: return "하차"
-        }
-    }
-    
-    var dateText: String? {
-        switch self {
-        case .watched: return "읽은 날짜"
-        case .watching: return "시작 날짜"
-        case .quit: return "종료 날짜"
-        }
-    }
-}
